+++ pysat/tests/classes/cls_instrument_property.py
@@ -17,6 +17,7 @@ import pytest
 import xarray as xr
 
 import pysat
+from pysat.utils import testing
 from pysat.utils.time import filter_datetime_input
 
 
@@ -38,14 +39,13 @@ class InstPropertyTests(object):
         """Check for error when instantiating with bad load keywords on init."""
 
         # Test that the correct error is raised
-        with pytest.raises(ValueError) as verr:
-            pysat.Instrument(platform=self.testInst.platform,
-                             name=self.testInst.name, num_samples=10,
-                             clean_level='clean',
-                             unsupported_keyword_yeah=True)
-
-        # Evaluate error message
-        assert str(verr).find("unknown keyword supplied") > 0
+        testing.eval_bad_input(pysat.Instrument, ValueError,
+                               "unknown keyword supplied",
+                               input_kwargs={'platform': self.testInst.platform,
+                                             'name': self.testInst.name,
+                                             'num_samples': 10,
+                                             'clean_level': 'clean',
+                                             'unsupported_keyword_yeah': True})
         return
 
     @pytest.mark.parametrize('kwarg', ['supported_tags', 'start', 'stop',
@@ -59,18 +59,17 @@ class InstPropertyTests(object):
             Name of reserved keyword.
 
         """
+        # Set the input
+        in_kwargs = {'platform': self.testInst.platform,
+                     'name': self.testInst.name, 'num_samples': 10,
+                     'clean_level': 'clean', kwarg: '1s'}
 
         # Check that the correct error is raised
-        with pytest.raises(ValueError) as err:
-            pysat.Instrument(platform=self.testInst.platform,
-                             name=self.testInst.name, num_samples=10,
-                             clean_level='clean',
-                             **{kwarg: '1s'})
-
-        # Check that the error message is correct
         estr = ''.join(('Reserved keyword "', kwarg, '" is not ',
                         'allowed at instantiation.'))
-        assert str(err).find(estr) >= 0
+        testing.eval_bad_input(pysat.Instrument, ValueError, estr,
+                               input_kwargs=in_kwargs)
+
         return
 
     @pytest.mark.parametrize('attr', ['_test_download', '_test_download_ci',
@@ -330,6 +329,7 @@ class InstPropertyTests(object):
 
         with pytest.raises(AttributeError) as aerr:
             self.testInst.bad_attr
+
         assert str(aerr).find("object has no attribute") >= 0
         return
 
@@ -468,11 +468,10 @@ class InstPropertyTests(object):
         import pysat.instruments.pysat_testing as test
         delattr(test, del_routine)
 
-        with pytest.raises(AttributeError) as aerr:
-            pysat.Instrument(inst_module=test, tag='',
-                             clean_level='clean')
         estr = 'A `{:}` function is required'.format(del_routine)
-        assert str(aerr).find(estr) >= 0
+        testing.eval_bad_input(pysat.Instrument, AttributeError, estr,
+                               input_kwargs={'inst_module': test, 'tag': '',
+                                             'clean_level': 'clean'})
         return
 
     @pytest.mark.parametrize("func, kwarg, val", [('init', 'test_init_kwarg',
@@ -602,13 +601,13 @@ class InstPropertyTests(object):
         self.testInst.kwargs['load']['undefined_keyword1'] = True
         self.testInst.kwargs['load']['undefined_keyword2'] = False
 
-        with pytest.raises(ValueError) as err:
-            # Instantiate instrument with new undefined keyword involved
-            eval(repr(self.testInst))
-
         estr = "".join(("unknown keywords supplied: ['undefined_keyword1',",
                         " 'undefined_keyword2']"))
-        assert str(err).find(estr) >= 0
+        with pytest.raises(ValueError) as verr:
+            eval(repr(self.testInst))
+
+        assert str(verr).find(estr) >= 0, "{:s} not found in {:}".format(
+            estr, str(verr))
         return
 
     def test_supported_input_keywords(self):
@@ -665,21 +664,18 @@ class InstPropertyTests(object):
             Text that should be contained in the error message.
 
         """
+        kwargs['platform'] = self.testInst.platform
+        kwargs['name'] = self.testInst.name
 
-        with pytest.raises(ValueError) as err:
-            pysat.Instrument(platform=self.testInst.platform,
-                             name=self.testInst.name,
-                             **kwargs)
-        assert str(err).find(estr) >= 0
+        testing.eval_bad_input(pysat.Instrument, ValueError, estr,
+                               input_kwargs=kwargs)
         return
 
     def test_get_var_type_code_unknown_type(self):
         """Ensure that Error is thrown if unknown type is supplied."""
 
-        with pytest.raises(TypeError) as err:
-            self.testInst._get_var_type_code(type(None))
-        estr = 'Unknown Variable'
-        assert str(err).find(estr) >= 0
+        testing.eval_bad_input(self.testInst._get_var_type_code, TypeError,
+                               'Unknown Variable', [type(None)])
         return
 
     @pytest.mark.parametrize("kwargs", [{'platform': 'doctor'},
@@ -711,27 +707,40 @@ class InstPropertyTests(object):
     def test_change_inst_pandas_format(self):
         """Test changing `pandas_format` attribute works."""
         new_format = not self.testInst.pandas_format
+        new_type = pds.DataFrame if new_format else xr.Dataset
 
-        # Current data format hidden attributes
+        # Save current data format hidden attributes
         current_null = self.testInst._null_data
         current_library = self.testInst._data_library
 
         # Assign inverted `pandas_format` setting
         self.testInst.pandas_format = new_format
 
-        # Confirm assignment
+        # Confirm assignment of visible and hidden attributes
         assert self.testInst.pandas_format == new_format
-
-        # Confirm that internal properties have changed
         assert not isinstance(self.testInst._null_data, type(current_null))
         assert current_library != self.testInst._data_library
 
         # Confirm internal consistency
         assert isinstance(self.testInst._null_data, self.testInst._data_library)
+        assert isinstance(self.testInst._null_data, new_type)
 
-        if new_format:
-            assert isinstance(self.testInst._null_data, pds.DataFrame)
-        else:
-            assert isinstance(self.testInst._null_data, xr.Dataset)
+        return
+
+    def test_change_inst_pandas_format_loaded_data(self):
+        """Test changing `pandas_format` attribute when data loaded."""
+
+        # Load data
+        self.testInst.load(date=self.ref_time)
+
+        # Get inverted pandas_format setting
+        new_format = not self.testInst.pandas_format
+
+        # Assign inverted `pandas_format` setting
+        with pytest.raises(ValueError) as err:
+            self.testInst.pandas_format = new_format
+
+        estr = "Can't change data type setting while data is "
+        assert str(err).find(estr) > 0
 
         return